requirements.txt
# Django Core
Django==6.0.1
djangorestframework==3.16.1

# JWT Authentication
djangorestframework-simplejwt==5.5.0

# CORS Headers
django-cors-headers==4.9.0

# Environment Variables
python-decouple==3.8

# YouTube Audio Download
yt-dlp==2025.01.26

# Audio Transcription (requires FFmpeg installed on system)
openai-whisper

# Quiz Generation (new SDK - https://github.com/googleapis/python-genai)
google-genai

# Additional Dependencies
python-dotenv==1.0.1
